refactor : 출력 변환기 생성
