+++ src/main/java/view/OutputView.java
@@ -1,15 +1,25 @@
 package view;
 
-import domain.*;
-
-import java.util.List;
-import java.util.stream.Collectors;
+import domain.Ladder;
+import domain.Line;
+import domain.Players;
 
 public class OutputView {
+    private final OutputFormatter outputFormatter;
+
+    public OutputView() {
+        this.outputFormatter = new OutputFormatter();
+    }
+
     public void printResult(Players rawPlayers, Ladder ladder) {
         System.out.println("실행결과");
         printNames(rawPlayers);
         printLadder(ladder);
+    }
+
+    private void printNames(Players players) {
+        String nameUnit = outputFormatter.toNameUnit(players);
+        System.out.println(nameUnit);
     }
 
     private void printLadder(Ladder ladder) {
@@ -17,35 +27,7 @@ public class OutputView {
     }
 
     private void printLine(Line rawLine) {
-        String line = rawLine.getPoints().stream()
-                .map(this::getStep)
-                .collect(Collectors.joining());
-        System.out.println("    " + line);
-    }
-
-    private String getStep(Point point) {
-        if (point.getStep().equals(Step.EXIST)) {
-            return "|" + "-----";
-        }
-        return "|" + "     ";
-    }
-
-    private void printNames(Players rawPlayers) {
-        List<Player> players = rawPlayers.getPlayers();
-        StringBuilder nameUnit = new StringBuilder();
-        for (Player player : players) {
-            nameUnit.append(getNameUnit(player));
-        }
-        System.out.println(nameUnit);
-    }
-
-    private String getNameUnit(Player player) {
-        String name = player.getName();
-        if (name.length() < 5) {
-            String leftBlank = " ".repeat(4 - name.length());
-            String rightBalnk = " ";
-            name = leftBlank + name + rightBalnk;
-        }
-        return name + " ";
+        String line = outputFormatter.toLine(rawLine);
+        System.out.println(line);
     }
 }
